+++ src/Module_1/part1.h
@@ -1,9 +1,9 @@
 #pragma once
 #include "part0.h"
 
-using u32 = unsigned long;
+using i64 = long long;
 
-char DIV_NN_Dk(const big_N& x, const big_N& y, u32 k); 
+char DIV_NN_Dk(const big_N& x, const big_N& y, i64 k); 
 
 big_N DIV_NN_N(const big_N& x, const big_N& y);
